Rework calculation of influence measures
The function `influence.ivreg()` now efficiently (re)uses matrix
factorisations, avoiding recalculations and potential numerical issues
due to explicit use of the inverse. The algorithm is not fully
iterative anymore. Most quantities are created in one go, with the
exception of diagonal matrix entries (which use `vapply()`). This
proved to be faster in testing.
The code now computes a QR of Z (for its hat values and residuals),
and one of X'PX (the matrix A). PX is taken directly from the input.
The order of operations wrt `denom` and `delta` is hopefully suitable
to avoid numeric issues. Calculation of `sigma` should now be a lot
faster and reuses appropriate quantities.
Benchmarks are favourable and `testthat` runs through.