Fixes #23
If a model is specified where the instruments perfectly predict all regressors,
a warning is now issued that no endogenous variables could be detected and that
all regressors appear to be exogenous.

Additionally, the documentation clarifies that fitting a model with only
a single right-hand side (i.e., treating all regressors as exogenous)
is possible for convenience, e.g., to facilitate model comparisons. This
now also works correctly for M and MM estimation.